requirements.txt
python-dotenv
jupyterlab
pandas
yfinance
pygwalker
openai
lightweight-charts
streamlit
streamlit-lightweight-charts